fix casts

+++ trunk/janus-uae/src/zfile_archive.cpp
@@ -1630,7 +1630,7 @@ static struct zfile *archive_access_adf (struct znode *zn)
 		for (i = 0; i < sfsblockcnt; i++)
 			bsize += sfsblocks[i].length * adf->blocksize;
 		if (bsize < size)
-			write_log (_T("SFS extracting error, %s size mismatch %d<%d\n"), z->name, bsize, size);
+			write_log (_T("SFS extracting error, %s size mismatch %d<%d\n"), z->name, (int)bsize, (int)size);
 
 		dst = z->data;
 		block = zn->offset;
